pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "\"Hexlet`s python project 'Pages analizator'\""
authors = ["ibrg <romanbezh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
